% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_corn.R
\docType{data}
\name{corn}
\alias{corn}
\title{Corn data}
\format{
A data frame with 34 rows and two columns.
\describe{
\item{method}{one of four different methods used to grow corn}
\item{outcome}{corn yield categorised into four levels}
}
}
\source{
Rayner and Best (2001)
}
\usage{
data(corn)
}
\description{
Corn yields when grown by four different methods.
}
\examples{
attach(corn)
KW(treatment = method, response = outcome)
}
\references{
Rayner, J.C.W. and Best, D.J. (2001). A Contingency Table Approach to Nonparametric Testing. Chapman & Hall/CRC: Boca Raton FL.
}
\keyword{corn}
\keyword{datasets}
