% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_intelligence.R
\docType{data}
\name{intelligence}
\alias{intelligence}
\title{Intelligence data}
\format{
A data frame with 15 rows and two columns.
\describe{
\item{age}{age of the respondents}
\item{score}{intelligence score achieved}
}
}
\source{
Rayner and Best (2001, section 8.1)
}
\usage{
data(intelligence)
}
\description{
Intelligence scores for individuals from different age groups.
}
\examples{
attach(intelligence)
gen_cor(x = rank(score), y = age, U = 2, V = 2)
}
\references{
Rayner, J. C. W. and Best, D. J. (2001). A Contingency Table Approach to Nonparametric Testing. Boca Raton: Chapman & Hall/CRC.
}
\keyword{datasets}
\keyword{intelligence}
