% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_KW.R
\name{KW}
\alias{KW}
\title{KW}
\usage{
KW(treatment, response)
}
\arguments{
\item{treatment}{a factor that fedines the group the response belongs to.}

\item{response}{a numeric variable measuring the response outcome.}
}
\value{
The function returns the test result of the Kruskal Wallis test with
adjustment for ties and without.
}
\description{
\code{KW} returns the test statistic and p-value for the Kruskal-Wallis test
adjusted for ties.
}
\details{
The Kruskal-Wallis test is a non-parametric equivalent to the one way ANOVA.
}
\examples{
attach(whiskey)
KW(treatment = maturity, response = grade)
}
\references{
Rayner, J.C.W and Livingston, G. C. (2022). An Introduction to Cochran-Mantel-Haenszel Testing and Nonparametric ANOVA. Wiley.
}
