% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_marriage.R
\docType{data}
\name{marriage}
\alias{marriage}
\title{Marriage data}
\format{
A data frame with 133 rows and three columns.
\describe{
\item{education}{highest level of eduction for the respondent}
\item{religion}{level of religiosity. Scores of 1, 2 and 3 are assigned to the religious categories fundamentalist, moderate and liberal respectively}
\item{opinion}{the response to the proposition: "Homosexuals should be able to marry". Scores of 1, 2 and 3 are assigned to the responses agree, neutral and
disagree respectively}
}
}
\source{
Agresti (2002)
}
\usage{
data(marriage)
}
\description{
Scores of 1, 2 and 3 are assigned to the responses agree, neutral and
disagree respectively to the proposition "Homosexuals should be able to
marry" and scores of 1, 2 and 3 are assigned to the religious categories
fundamentalist, moderate and liberal respectively.
}
\examples{
attach(marriage)
lm(as.numeric(opinion)~religion+education)
}
\references{
Agresti, A. (2002). Categorical Data Analysis, 2nd ed; Wiley: New York.
}
\keyword{datasets}
\keyword{marriage}
