% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mu_r.R
\name{mu_r}
\alias{mu_r}
\title{mu_r}
\usage{
mu_r(x, r)
}
\arguments{
\item{x}{a numeric vector.}

\item{r}{the degree moment requiring calculation}
}
\value{
Returns the estimated rth moment of a vector.
}
\description{
\code{mu_r} returns the estimated rth moment of a vector.
}
\examples{
x_vec = rnorm(10)
mu_r(x_vec,2)
}
\references{
Rayner, J.C.W and Livingston, G. C. (2022). An Introduction to Cochran-Mantel-Haenszel Testing and Nonparametric ANOVA. Wiley.
}
