% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_orthogonal_scores.R
\name{orthogonal_scores}
\alias{orthogonal_scores}
\title{orthogonal_scores}
\usage{
orthogonal_scores(x, degree, n_strata = 1)
}
\arguments{
\item{x}{a vector of scores, either a factor or numeric.}

\item{degree}{the degree of orthogonal scores required.}

\item{n_strata}{optional argument for indicating the number of strata to
apply the scores to.}
}
\value{
Returns a matrix of orthogonal scores.
}
\description{
\code{orthogonal_scores} returns orthogonal scores weighted by prevalence in
the data.
}
\examples{
attach(jam)
orthogonal_scores(x = sweetness, degree = 2, n_strata = 8)
}
\references{
Rayner, J.C.W and Livingston, G. C. (2022). An Introduction to Cochran-Mantel-Haenszel Testing and Nonparametric ANOVA. Wiley.
}
