% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_peanuts.R
\docType{data}
\name{peanuts}
\alias{peanuts}
\title{Peanuts data}
\format{
A data frame with 16 rows and four columns.
\describe{
\item{yield}{the peanut yield}
\item{treatment}{the variety of peanut plant}
\item{row}{the row the plants were grown in of the plot}
\item{col}{the column the plants were grown in of the plot}
}
}
\source{
http://www.math.montana.edu/jobo/st541/sec3c.pdf
}
\usage{
data(peanuts)
}
\description{
A plant biologist conducted an experiment to compare the yields of four
varieties of peanuts, denoted as A, B, C, and D. A plot of land was divided
into 16 subplots (four rows and four columns).
}
\examples{
attach(peanuts)
CARL(y = yield, treatment = treatment, block1 = row, block2 = col)
}
\references{
Rayner, J.C.W and Livingston, G. C. (2022). An Introduction to Cochran-Mantel-Haenszel Testing and Nonparametric ANOVA. Wiley.
}
\keyword{LSD}
\keyword{datasets}
\keyword{peanuts}
