% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_Poly.R
\name{Poly}
\alias{Poly}
\title{Poly}
\usage{
Poly(x, p)
}
\arguments{
\item{x}{a vector of numeric scores.}

\item{p}{a vector of weights corresponding to the elements of x.}
}
\value{
Returns a matrix of orthomornal scores based on the weights provided.
}
\description{
\code{Poly} returns t-1 orthonormal scores weighted by a weights parameter.
The function uses Emerson Recursion.
}
\examples{
x = 1:5
p = rep(0.2,5)
Poly(x = x, p = p)
}
\references{
Rayner, J.C.W and Livingston, G. C. (2022). An Introduction to Cochran-Mantel-Haenszel Testing and Nonparametric ANOVA. Wiley.
Rayner, J.C.W., Thas, O. and De Boeck, B. (2008), A GENERALIZED EMERSON RECURRENCE RELATION. Australian & New Zealand Journal of Statistics, 50: 235-240.
}
